+++ client/src/shared/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './axios';
 export * from './toastify';
 export * from './interface';
+export * from './date';
